feat: v0.7.0 desktop polish — toolbar cleanup, library refactor
- vm-window: drop the duplicate VM-name rendering. `windowToolbarStyle(.unified(showsTitle: false))` strips both the left-aligned title and the centered capsule drawn by the unified title bar; the `.principal` ToolbarItem remains as the single centered VM name.
- fullscreen: replace the in-view `FullscreenChrome` overlay with native macOS behavior. On macOS 15+ apply `windowToolbarFullScreenVisibility(.onHover)` so the toolbar auto-hides and reveals on cursor-to-top (same gesture that unfurls the menu bar). On macOS 14 fall back to `LegacyFullscreenConfigurator` (renamed from `ImmersiveFullscreenConfigurator`) driving `NSApp.presentationOptions` on the enter/exit fullscreen notifications.
- library: extract `HostStatusRibbon`, `LibraryEmptyStates`, `VMCard`, `VMListView` out of the monolithic `LibraryView` (~800 LOC moved out).
- supporting work across Lumina core (`VM`, `CommandRunner`, `MacOSVM`), Theme, AppModel, LuminaDesktopSession, DesktopOSCatalog, and CI workflows.
