+++ Sources/LuminaDesktopKit/LibraryView.swift
@@ -3,6 +3,14 @@
 // v0.7.0 M6 — Mac-native rewrite. NavigationSplitView + materials +
 // proper sidebar + detail pane. Phosphor-amber accent on a translucent
 // dark surface that picks up the desktop wallpaper through the window.
+//
+// Split across files for locatability:
+//   - `HostStatusRibbon.swift`     — top-of-detail metric strip
+//   - `LibraryEmptyStates.swift`   — empty-library hero + filter empties
+//   - `VMCard.swift`               — grid card + state chip + preview + action + distro chip
+//   - `VMListView.swift`           — dense-row list layout
+//   - `Heartbeat.swift`            — pulse indicator used by card and row
+//   - this file: top-level library composition + toolbar chrome
 
 import SwiftUI
 import LuminaBootable
@@ -82,8 +90,11 @@ public struct LibraryView: View {
             detail
         }
         .navigationSplitViewStyle(.balanced)
+        // Instant scheme swap — native macOS behavior (Mail, Safari, Xcode).
+        // The prior `.animation(.easeInOut(duration: 0.18), value: appearanceState)`
+        // cross-faded the whole tree over 180ms, which reads as lag. Let
+        // SwiftUI handle color scheme switching without an explicit animation.
         .preferredColorScheme(appearanceState.colorScheme)
-        .animation(.easeInOut(duration: 0.18), value: appearanceState)
         .frame(minWidth: 1080, minHeight: 660)
         .background(MaterialBackground(material: .underWindowBackground))
         .luminaWindowChrome()
@@ -285,69 +296,6 @@ public struct LibraryView: View {
     }
 }
 
-// ── HOST STATUS RIBBON ────────────────────────────────────────────
-/// Thin strip at the top of the detail pane. Always shows host
-/// constraints (model / RAM / free disk) + running-VM count. This fills
-/// the library's breathing room with *useful* information — Victor's
-/// "make the invisible visible" applied to the host/guest contract.
-@MainActor
-public struct HostStatusRibbon: View {
-    @Bindable var model: AppModel
-
-    public var body: some View {
-        let host = HostInfo.current
-        let freeDisk = HostInfo.freeDiskBytes(at: model.store.rootURL) ?? 0
-        let running = model.sessions.values.filter { $0.status.isLive }.count
-        let totalUsed = model.bundles.reduce(0) { $0 + $1.actualDiskBytes }
-
-        HStack(spacing: 16) {
-            metric(label: "HOST", value: host.modelName.uppercased())
-            divider
-            metric(label: "RAM", value: formatBytesHuman(host.physicalMemoryBytes))
-            metric(label: "CORES", value: "\(host.processorCount)")
-            metric(label: "LIBRARY FREE", value: formatBytesHuman(freeDisk),
-                   warn: freeDisk < 16 * 1024 * 1024 * 1024)
-            divider
-            metric(label: "VMS", value: "\(model.bundles.count)")
-            metric(label: "USED", value: formatBytesHuman(totalUsed))
-            metric(label: "RUNNING", value: "\(running)",
-                   accent: running > 0 ? LuminaTheme.ok : LuminaTheme.inkDim)
-            Spacer()
-        }
-        .padding(.horizontal, 20)
-        .padding(.vertical, 10)
-        .background(
-            Rectangle().fill(LuminaTheme.bg2.opacity(0.5))
-        )
-        .overlay(
-            Rectangle().fill(LuminaTheme.rule).frame(height: 1),
-            alignment: .bottom
-        )
-    }
-
-    @ViewBuilder
-    private func metric(label: String, value: String,
-                        accent: Color = LuminaTheme.ink,
-                        warn: Bool = false) -> some View {
-        HStack(spacing: 6) {
-            Text(label)
-                .font(.system(size: 9, weight: .medium, design: .monospaced))
-                .tracking(1.2)
-                .foregroundStyle(LuminaTheme.inkMute)
-            Text(value)
-                .font(.system(size: 11, weight: .medium, design: .monospaced))
-                .foregroundStyle(warn ? LuminaTheme.warn : accent)
-                .monospacedDigit()
-        }
-    }
-
-    private var divider: some View {
-        Rectangle()
-            .fill(LuminaTheme.rule2)
-            .frame(width: 1, height: 14)
-    }
-}
-
 // ── LAYOUT PICKER ─────────────────────────────────────────────────
 @MainActor
 public struct LayoutPicker: View {
@@ -355,6 +303,11 @@ public struct LayoutPicker: View {
     private var current: LibraryLayout {
         LibraryLayout(rawValue: layoutRaw) ?? .list
     }
+
+    public init(layoutRaw: Binding<String>) {
+        _layoutRaw = layoutRaw
+    }
+
     public var body: some View {
         Picker("Layout", selection: $layoutRaw) {
             ForEach(LibraryLayout.allCases, id: \.rawValue) { l in
@@ -372,6 +325,11 @@ public struct LayoutPicker: View {
 public struct AppearanceMenu: View {
     @Binding var current: AppearancePreference
     @Binding var persistedRaw: String
+
+    public init(current: Binding<AppearancePreference>, persistedRaw: Binding<String>) {
+        _current = current
+        _persistedRaw = persistedRaw
+    }
 
     public var body: some View {
         // Fast path: flip the @State (triggers instant redraw) + persist
@@ -407,10 +365,11 @@ public struct AppearanceMenu: View {
 
     private func setAppearance(_ next: AppearancePreference) {
         guard current != next else { return }
-        // Flip @State first — this is what drives the visible redraw.
-        withAnimation(.easeInOut(duration: 0.14)) {
-            current = next
-        }
+        // Instant flip — no animation. macOS appearance changes are
+        // native-instant in Apple apps; wrapping in `withAnimation`
+        // cross-faded the tree and read as lag. The @State write drives
+        // the visible redraw in the current runloop tick.
+        current = next
         // Persist in the next runloop tick so UserDefaults' synchronous
         // willSet chain doesn't block the paint.
         DispatchQueue.main.async {
@@ -454,725 +413,5 @@ public struct LuminaSearchField: View {
             RoundedRectangle(cornerRadius: 6)
                 .stroke(LuminaTheme.rule2, lineWidth: 0.5)
         )
-    }
-}
-
-// ── EMPTY STATES ──────────────────────────────────────────────────
-@MainActor
-public struct EmptyStateView: View {
-    let onChoose: (String?) -> Void   // nil = open blank wizard, else pre-pick tile id
-
-    public var body: some View {
-        VStack(spacing: 0) {
-            Spacer()
-            VStack(spacing: 28) {
-                BrandMarkLarge()
-                    .frame(width: 80, height: 80)
-                VStack(spacing: 8) {
-                    Text("subprocess.run()")
-                        .foregroundStyle(LuminaTheme.accent)
-                        .font(.system(size: 36, weight: .medium, design: .monospaced))
-                        .tracking(-0.5)
-
-                    HStack(spacing: 10) {
-                        Text("for")
-                            .font(.system(size: 36, weight: .medium, design: .monospaced))
-                            .foregroundStyle(LuminaTheme.ink)
-                            .tracking(-0.5)
-                        Text("virtual machines.")
-                            .font(.system(.title, design: .serif).italic())
-                            .foregroundStyle(LuminaTheme.inkDim)
-                    }
-                }
-                Text("Spin up Ubuntu, Kali, Windows 11 ARM, or macOS — \nand throw it away when you're done.")
-                    .font(.system(size: 14))
-                    .foregroundStyle(LuminaTheme.inkDim)
-                    .multilineTextAlignment(.center)
-                    .lineSpacing(4)
-
-                HStack(spacing: 10) {
-                    PrimaryAction(label: "Try Ubuntu", systemImage: "circle.hexagongrid.fill", isPrimary: true) {
-                        onChoose("ubuntu-24.04")
-                    }
-                    PrimaryAction(label: "Install Windows 11", systemImage: "macwindow") {
-                        onChoose("windows-11-arm")
-                    }
-                    PrimaryAction(label: "Install macOS", systemImage: "apple.logo") {
-                        onChoose("macos-latest")
-                    }
-                    PrimaryAction(label: "Use my own…", systemImage: "doc.badge.plus") {
-                        onChoose("byo-file")
-                    }
-                }
-                .padding(.top, 8)
-            }
-            Spacer()
-            HStack(spacing: 14) {
-                Image(systemName: "arrow.down.doc")
-                    .foregroundStyle(LuminaTheme.inkMute)
-                Text("Drop an ISO or IPSW anywhere on this window")
-                    .font(.system(size: 12))
-                    .foregroundStyle(LuminaTheme.inkMute)
-            }
-            .padding(.bottom, 28)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-}
-
-@MainActor
-public struct EmptyFilterView: View {
-    let section: SidebarSection
-    public var body: some View {
-        VStack(spacing: 12) {
-            Image(systemName: section.systemImage)
-                .font(.system(size: 32, weight: .light))
-                .foregroundStyle(LuminaTheme.inkMute)
-            Text("No VMs in \(section.rawValue)")
-                .font(.system(size: 15, weight: .medium))
-                .foregroundStyle(LuminaTheme.ink)
-            Text("Create one from the toolbar.")
-                .font(.system(size: 12))
-                .foregroundStyle(LuminaTheme.inkDim)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-}
-
-@MainActor
-public struct BrandMarkLarge: View {
-    public init() {}
-    public var body: some View {
-        ZStack {
-            // back square (cream)
-            RoundedRectangle(cornerRadius: 8)
-                .stroke(LuminaTheme.ink.opacity(0.4), lineWidth: 2)
-                .frame(width: 56, height: 56)
-                .offset(x: 12, y: 12)
-            // front square (amber, with subtle glow)
-            RoundedRectangle(cornerRadius: 8)
-                .stroke(LuminaTheme.accent, lineWidth: 2)
-                .frame(width: 56, height: 56)
-                .shadow(color: LuminaTheme.accent.opacity(0.4), radius: 8)
-        }
-    }
-}
-
-@MainActor
-public struct PrimaryAction: View {
-    let label: String
-    let systemImage: String
-    let isPrimary: Bool
-    let action: () -> Void
-    @State private var hovering = false
-
-    public init(label: String, systemImage: String, isPrimary: Bool = false, action: @escaping () -> Void) {
-        self.label = label
-        self.systemImage = systemImage
-        self.isPrimary = isPrimary
-        self.action = action
-    }
-
-    public var body: some View {
-        Button(action: action) {
-            HStack(spacing: 7) {
-                Image(systemName: systemImage)
-                    .font(.system(size: 12, weight: .medium))
-                Text(label)
-                    .font(.system(size: 13, weight: isPrimary ? .semibold : .regular))
-            }
-            .foregroundStyle(isPrimary ? Color.black : (hovering ? LuminaTheme.accent : LuminaTheme.ink))
-            .padding(.horizontal, 14)
-            .padding(.vertical, 8)
-            .background(
-                RoundedRectangle(cornerRadius: 7)
-                    .fill(isPrimary ? LuminaTheme.accent : LuminaTheme.bg1.opacity(hovering ? 0.9 : 0.5))
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 7)
-                    .stroke(isPrimary ? Color.clear : (hovering ? LuminaTheme.accent.opacity(0.5) : LuminaTheme.rule2),
-                            lineWidth: 1)
-            )
-        }
-        .buttonStyle(.plain)
-        .onHover { hovering = $0 }
-        .animation(.easeOut(duration: 0.12), value: hovering)
-    }
-}
-
-// ── GRID ──────────────────────────────────────────────────────────
-// ── STATE SHARED BY CARD + ROW ─────────────────────────────────────
-/// Active (colored) state chip. Returns nil for `.stopped` — per Victor's
-/// rule, idle gets no chip. The eye should be drawn to states that
-/// demand response, not to "exists".
-@MainActor
-struct StateChip: View {
-    let status: LuminaDesktopSession.Status
-
-    var body: some View {
-        if let (label, color, pulse) = chipSpec {
-            HStack(spacing: 5) {
-                Circle()
-                    .fill(color)
-                    .frame(width: 6, height: 6)
-                    .shadow(color: pulse ? color.opacity(0.7) : .clear, radius: 4)
-                Text(label)
-                    .font(.system(size: 10, weight: .medium, design: .monospaced))
-                    .tracking(0.5)
-                    .textCase(.uppercase)
-                    .foregroundStyle(color)
-            }
-            .padding(.horizontal, 7)
-            .padding(.vertical, 3)
-            .overlay(
-                Capsule().stroke(color.opacity(0.35), lineWidth: 0.5)
-            )
-        }
-    }
-
-    private var chipSpec: (label: String, color: Color, pulse: Bool)? {
-        switch status {
-        case .running:      ("RUNNING",  LuminaTheme.ok,  true)
-        case .booting:      ("BOOTING",  LuminaTheme.accent, true)
-        case .paused:       ("PAUSED",   LuminaTheme.warn, false)
-        case .crashed:      ("CRASHED",  LuminaTheme.err, false)
-        case .shuttingDown: ("STOPPING", LuminaTheme.inkDim, false)
-        case .stopped:      nil   // Idle = silence
-        }
-    }
-}
-
-/// Terminal-style preview of a VM's current state — the card's primary
-/// content. Replaces the decorative centered glyph. Shows what's
-/// actually going on: last-boot time, disk usage, snapshot count,
-/// install state if the VM hasn't been booted yet.
-@MainActor
-struct VMStatePreview: View {
-    let bundle: VMBundle
-    let status: LuminaDesktopSession.Status
-    let bootedAt: Date?
-    let density: Density
-
-    enum Density { case card, row }
-
-    private var isInstalling: Bool {
-        // No lastBootedAt + a pending ISO sidecar present = first boot incoming.
-        if bundle.manifest.lastBootedAt != nil { return false }
-        let sidecar = bundle.rootURL.appendingPathComponent("pending-iso.path")
-        return FileManager.default.fileExists(atPath: sidecar.path)
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: density == .card ? 4 : 2) {
-            ForEach(lines, id: \.self) { line in
-                Text(line)
-                    .font(.system(size: density == .card ? 11 : 10, design: .monospaced))
-                    .foregroundStyle(LuminaTheme.inkDim)
-                    .lineLimit(1)
-            }
-        }
-    }
-
-    private var lines: [String] {
-        if isInstalling {
-            return [
-                "› stage  installer attached",
-                "› disk   0 of \(bundle.diskCapFormatted)",
-                "› awaiting first boot",
-            ]
-        }
-        switch status {
-        case .running:
-            return [
-                "› up     just now",
-                "› disk   \(bundle.diskUsedFormatted) / \(bundle.diskCapFormatted)",
-                "› snaps  \(bundle.snapshotCount)",
-            ]
-        case .booting:
-            return [
-                "› vm.boot() in flight",
-                "› virtio devices initialising",
-                "› disk   \(bundle.diskUsedFormatted) / \(bundle.diskCapFormatted)",
-            ]
-        case .crashed:
-            return [
-                "› vm_crashed",
-                "› last   \(bundle.lastBootedRelative)",
-                "› logs   available",
-            ]
-        case .stopped, .shuttingDown, .paused:
-            return [
-                "› last   \(bundle.lastBootedRelative)",
-                "› disk   \(bundle.diskUsedFormatted) / \(bundle.diskCapFormatted)",
-                "› snaps  \(bundle.snapshotCount)",
-            ]
-        }
-    }
-}
-
-/// Explicit action on the card/row: ▶ Boot / ■ Stop. Appears always in
-/// list view; hover-only in grid view. Direct manipulation — the verb
-/// is visible, not hidden behind a navigation step.
-@MainActor
-struct VMActionButton: View {
-    @Bindable var session: LuminaDesktopSession
-    let compact: Bool
-
-    @Environment(\.openWindow) private var openWindow
-    @State private var hovering = false
-
-    var body: some View {
-        Button {
-            switch session.status {
-            case .stopped, .crashed:
-                // Open the VM window immediately so the user sees the
-                // booting screen → framebuffer handoff, then kick off
-                // the boot. Without this the VM boots headless and the
-                // user has to click the card to see anything.
-                openWindow(id: "vm-window", value: session.bundle.manifest.id)
-                Task { await session.boot() }
-            case .running, .paused:
-                Task { await session.shutdown() }
-            case .booting, .shuttingDown:
-                break
-            }
-        } label: {
-            HStack(spacing: 5) {
-                Image(systemName: verb.glyph)
-                    .font(.system(size: compact ? 10 : 11, weight: .semibold))
-                if !compact {
-                    Text(verb.label)
-                        .font(.system(size: 11, weight: .semibold, design: .monospaced))
-                        .tracking(0.5)
-                }
-            }
-            .foregroundStyle(verb.foreground(hovering: hovering))
-            .padding(.horizontal, compact ? 8 : 10)
-            .padding(.vertical, compact ? 5 : 6)
-            .background(
-                Capsule().fill(verb.background(hovering: hovering))
-            )
-            .overlay(
-                Capsule().stroke(verb.stroke(hovering: hovering), lineWidth: 1)
-            )
-        }
-        .buttonStyle(.plain)
-        .onHover { hovering = $0 }
-        .help(verb.tooltip)
-        .disabled(verb.disabled)
-    }
-
-    private enum Verb {
-        case boot, stop, bootingNow, stoppingNow, restart
-
-        var label: String {
-            switch self {
-            case .boot: "BOOT"
-            case .stop: "STOP"
-            case .bootingNow: "BOOTING"
-            case .stoppingNow: "STOPPING"
-            case .restart: "RESTART"
-            }
-        }
-
-        var glyph: String {
-            switch self {
-            case .boot: "play.fill"
-            case .stop: "stop.fill"
-            case .bootingNow, .stoppingNow: "circle.dashed"
-            case .restart: "arrow.counterclockwise"
-            }
-        }
-
-        var disabled: Bool {
-            self == .bootingNow || self == .stoppingNow
-        }
-
-        var tooltip: String {
-            switch self {
-            case .boot: "Boot this VM"
-            case .stop: "Shut down this VM"
-            case .bootingNow: "Booting in progress…"
-            case .stoppingNow: "Shutting down…"
-            case .restart: "Restart this VM"
-            }
-        }
-
-        func foreground(hovering: Bool) -> Color {
-            switch self {
-            case .boot: hovering ? Color.black : LuminaTheme.accent
-            case .stop: hovering ? Color.white : LuminaTheme.err
-            default: LuminaTheme.inkMute
-            }
-        }
-
-        func background(hovering: Bool) -> Color {
-            switch self {
-            case .boot: hovering ? LuminaTheme.accent : LuminaTheme.accent.opacity(0.1)
-            case .stop: hovering ? LuminaTheme.err : LuminaTheme.err.opacity(0.1)
-            default: Color.clear
-            }
-        }
-
-        func stroke(hovering: Bool) -> Color {
-            switch self {
-            case .boot: hovering ? LuminaTheme.accent : LuminaTheme.accent.opacity(0.5)
-            case .stop: hovering ? LuminaTheme.err : LuminaTheme.err.opacity(0.5)
-            default: LuminaTheme.rule2
-            }
-        }
-    }
-
-    private var verb: Verb {
-        switch session.status {
-        case .stopped, .crashed: .boot
-        case .running, .paused: .stop
-        case .booting: .bootingNow
-        case .shuttingDown: .stoppingNow
-        }
-    }
-}
-
-/// Left-edge OS stripe — 3pt vertical strip. The card/row body stays
-/// neutral; identity comes from the stripe, not the background.
-struct OSStripe: View {
-    let family: OSFamily
-    let height: CGFloat
-    var body: some View {
-        Rectangle()
-            .fill(LuminaTheme.osStripe(family.rawValue))
-            .frame(width: 3)
-            .frame(maxHeight: height > 0 ? height : .infinity)
-    }
-}
-
-// ── GRID VIEW ─────────────────────────────────────────────────────
-@MainActor
-public struct VMGridView: View {
-    @Bindable var model: AppModel
-    let bundles: [VMBundle]
-    @Environment(\.openWindow) private var openWindow
-
-    private let columns = [
-        GridItem(.adaptive(minimum: 300, maximum: 380), spacing: 14)
-    ]
-
-    public var body: some View {
-        ScrollView {
-            LazyVGrid(columns: columns, spacing: 14) {
-                ForEach(bundles, id: \.manifest.id) { bundle in
-                    VMCard(model: model, bundle: bundle,
-                           openWindow: { openWindow(id: "vm-window", value: bundle.manifest.id) })
-                }
-            }
-            .padding(20)
-        }
-    }
-}
-
-@MainActor
-public struct VMCard: View {
-    @Bindable var model: AppModel
-    let bundle: VMBundle
-    let openWindow: () -> Void
-    @State private var isHovering = false
-    @State private var stats: VMLiveStats
-
-    init(model: AppModel, bundle: VMBundle, openWindow: @escaping () -> Void) {
-        self.model = model
-        self.bundle = bundle
-        self.openWindow = openWindow
-        _stats = State(initialValue: VMLiveStats(bundle: bundle))
-    }
-
-    private var session: LuminaDesktopSession {
-        model.session(for: bundle)
-    }
-
-    public var body: some View {
-        let brand = OSBranding.brand(for: bundle.manifest.osVariant,
-                                     family: bundle.manifest.osFamily)
-        HStack(spacing: 0) {
-            // Per-OS brand stripe (thicker than the default 3pt for impact)
-            Rectangle()
-                .fill(brand.accent)
-                .frame(width: 4)
-                .shadow(color: isHovering ? brand.accent.opacity(0.4) : .clear,
-                        radius: isHovering ? 4 : 0)
-            content(brand: brand)
-        }
-        .background(
-            ZStack {
-                // Base plate
-                RoundedRectangle(cornerRadius: 8)
-                    .fill(LuminaTheme.bg1.opacity(isHovering ? 0.82 : 0.62))
-                // Per-OS subtle tint gradient top→bottom
-                RoundedRectangle(cornerRadius: 8)
-                    .fill(
-                        LinearGradient(
-                            colors: brand.bodyGradient,
-                            startPoint: .top, endPoint: .bottom
-                        )
-                    )
-                    .opacity(isHovering ? 1.0 : 0.7)
-            }
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 8)
-                .stroke(isHovering ? brand.accent.opacity(0.7) : LuminaTheme.rule,
-                        lineWidth: isHovering ? 1 : 1)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 8))
-        .animation(.easeOut(duration: 0.14), value: isHovering)
-        .onHover { isHovering = $0 }
-        .onTapGesture { openWindow() }
-        .contextMenu {
-            Button("Open") { openWindow() }
-            Button("Reveal in Finder") {
-                NSWorkspace.shared.activateFileViewerSelecting([bundle.rootURL])
-            }
-            Divider()
-            Button("Move to Trash", role: .destructive) {
-                model.deleteBundle(bundle)
-            }
-        }
-    }
-
-    private func content(brand: OSBrand) -> some View {
-        VStack(alignment: .leading, spacing: 10) {
-            // Header: distro chip + VM name + state pill
-            HStack(alignment: .firstTextBaseline, spacing: 8) {
-                if session.status == .running {
-                    Heartbeat(color: LuminaTheme.ok)
-                }
-                Text(bundle.manifest.name)
-                    .font(.system(size: 14, weight: .semibold))
-                    .foregroundStyle(LuminaTheme.ink)
-                    .lineLimit(1)
-                Spacer()
-                StateChip(status: session.status)
-            }
-
-            // Distro chip — distinctive brand badge
-            DistroChip(brand: brand)
-
-            VMStatePreview(bundle: bundle, status: session.status,
-                           bootedAt: bundle.manifest.lastBootedAt,
-                           density: .card)
-            DiskSparkline(stats: stats,
-                          tint: brand.accent,
-                          running: session.status == .running)
-                .frame(height: 24)
-            HStack {
-                Text(bundle.manifest.osVariant)
-                    .font(.system(size: 10, weight: .medium, design: .monospaced))
-                    .tracking(0.4)
-                    .textCase(.uppercase)
-                    .foregroundStyle(LuminaTheme.inkMute)
-                Spacer()
-                if isHovering {
-                    VMActionButton(session: session, compact: false)
-                        .transition(.opacity.combined(with: .move(edge: .trailing)))
-                }
-            }
-            .animation(.easeOut(duration: 0.15), value: isHovering)
-        }
-        .padding(14)
-    }
-}
-
-/// Distro badge — small pill with the OS's glyph + its own brand name
-/// in its own typographic style. Ubuntu's lowercase serif, Kali's bold
-/// caps, macOS's sans-semibold. At-a-glance recognition.
-@MainActor
-struct DistroChip: View {
-    let brand: OSBrand
-
-    var body: some View {
-        HStack(spacing: 6) {
-            Image(systemName: brand.glyph)
-                .font(.system(size: 12, weight: .semibold))
-                .foregroundStyle(brand.textOnAccent)
-            Text(brand.displayName)
-                .font(OSBranding.nameFont(for: brand))
-                .foregroundStyle(brand.textOnAccent)
-        }
-        .padding(.horizontal, 10)
-        .padding(.vertical, 5)
-        .background(
-            Capsule().fill(brand.accent)
-        )
-    }
-}
-
-// ── LIST VIEW (default) ───────────────────────────────────────────
-@MainActor
-public struct VMListView: View {
-    @Bindable var model: AppModel
-    let bundles: [VMBundle]
-    @Environment(\.openWindow) private var openWindow
-
-    public var body: some View {
-        ScrollView {
-            VStack(spacing: 4) {
-                // Column header — dense like Activity Monitor / Linear / `lumina ps`
-                HStack(spacing: 12) {
-                    columnHeader("NAME", width: nil, align: .leading)
-                    columnHeader("STATE", width: 80)
-                    columnHeader("LAST BOOT", width: 100)
-                    columnHeader("DISK", width: 120)
-                    columnHeader("SNAPS", width: 50)
-                    columnHeader("", width: 72)
-                }
-                .padding(.horizontal, 14)
-                .padding(.vertical, 7)
-                .background(
-                    Rectangle().fill(LuminaTheme.bg2.opacity(0.5))
-                )
-                .overlay(
-                    Rectangle().fill(LuminaTheme.rule).frame(height: 1),
-                    alignment: .bottom
-                )
-
-                LazyVStack(spacing: 1) {
-                    ForEach(bundles, id: \.manifest.id) { bundle in
-                        VMRow(model: model, bundle: bundle)
-                            .onTapGesture {
-                                openWindow(id: "vm-window", value: bundle.manifest.id)
-                            }
-                    }
-                }
-            }
-            .padding(.horizontal, 16)
-            .padding(.top, 12)
-            .padding(.bottom, 20)
-        }
-    }
-
-    private func columnHeader(_ label: String, width: CGFloat?, align: Alignment = .leading) -> some View {
-        Text(label)
-            .font(.system(size: 10, weight: .medium, design: .monospaced))
-            .tracking(1.2)
-            .foregroundStyle(LuminaTheme.inkMute)
-            .frame(width: width, alignment: align)
-            .frame(maxWidth: width == nil ? .infinity : nil, alignment: align)
-    }
-}
-
-@MainActor
-public struct VMRow: View {
-    @Bindable var model: AppModel
-    let bundle: VMBundle
-    @Environment(\.openWindow) private var openWindow
-    @State private var isHovering = false
-    @State private var stats: VMLiveStats
-
-    init(model: AppModel, bundle: VMBundle) {
-        self.model = model
-        self.bundle = bundle
-        _stats = State(initialValue: VMLiveStats(bundle: bundle))
-    }
-
-    private var session: LuminaDesktopSession {
-        model.session(for: bundle)
-    }
-
-    public var body: some View {
-        let brand = OSBranding.brand(for: bundle.manifest.osVariant,
-                                     family: bundle.manifest.osFamily)
-        HStack(spacing: 0) {
-            Rectangle()
-                .fill(brand.accent)
-                .frame(width: 3)
-                .frame(maxHeight: 44)
-            HStack(spacing: 12) {
-                // NAME + live heartbeat when running + distro glyph
-                HStack(spacing: 10) {
-                    Image(systemName: brand.glyph)
-                        .font(.system(size: 14))
-                        .foregroundStyle(brand.accent)
-                        .frame(width: 20)
-                    VStack(alignment: .leading, spacing: 2) {
-                        HStack(spacing: 6) {
-                            if session.status == .running {
-                                Heartbeat(color: LuminaTheme.ok)
-                            }
-                            Text(bundle.manifest.name)
-                                .font(.system(size: 13, weight: .medium))
-                                .foregroundStyle(LuminaTheme.ink)
-                                .lineLimit(1)
-                        }
-                        Text(bundle.manifest.osVariant)
-                            .font(.system(size: 10, design: .monospaced))
-                            .foregroundStyle(LuminaTheme.inkMute)
-                            .lineLimit(1)
-                    }
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-                // STATE
-                HStack {
-                    StateChip(status: session.status)
-                    Spacer()
-                }
-                .frame(width: 80)
-
-                // LAST BOOT
-                Text(bundle.lastBootedRelative)
-                    .font(.system(size: 11, design: .monospaced))
-                    .foregroundStyle(LuminaTheme.inkDim)
-                    .frame(width: 100, alignment: .leading)
-
-                // DISK — text + live sparkline (per-OS brand accent)
-                VStack(alignment: .leading, spacing: 3) {
-                    Text("\(bundle.diskUsedFormatted) / \(bundle.diskCapFormatted)")
-                        .font(.system(size: 11, design: .monospaced))
-                        .foregroundStyle(LuminaTheme.inkDim)
-                        .monospacedDigit()
-                    DiskSparkline(stats: stats,
-                                  tint: brand.accent,
-                                  running: session.status == .running)
-                        .frame(height: 12)
-                }
-                .frame(width: 120, alignment: .leading)
-
-                // SNAPS
-                Text("\(bundle.snapshotCount)")
-                    .font(.system(size: 11, design: .monospaced))
-                    .foregroundStyle(LuminaTheme.inkDim)
-                    .frame(width: 50, alignment: .leading)
-                    .monospacedDigit()
-
-                // ACTION (always visible in list — this is the working view)
-                VMActionButton(session: session, compact: true)
-                    .frame(width: 72, alignment: .trailing)
-            }
-            .padding(.horizontal, 14)
-            .padding(.vertical, 9)
-        }
-        .background(
-            Rectangle()
-                .fill(isHovering ? LuminaTheme.bg1.opacity(0.75) : Color.clear)
-        )
-        .overlay(
-            Rectangle()
-                .fill(LuminaTheme.rule2.opacity(0.5))
-                .frame(height: 1),
-            alignment: .bottom
-        )
-        .animation(.easeOut(duration: 0.08), value: isHovering)
-        .onHover { isHovering = $0 }
-        .contextMenu {
-            Button("Open") {
-                openWindow(id: "vm-window", value: bundle.manifest.id)
-            }
-            Button("Reveal in Finder") {
-                NSWorkspace.shared.activateFileViewerSelecting([bundle.rootURL])
-            }
-            Divider()
-            Button("Move to Trash", role: .destructive) {
-                model.deleteBundle(bundle)
-            }
-        }
     }
 }
